refactor(kv-store): move store identity into sqlite-opfs and drop the lmdb-v2 re-export
The identity slug only has one consumer outside the browser backend: the pxe node-side helper,
whose lmdb directory names never interoperate with browser DB names. Let the helper compose its
own directory-name format inline and keep the identity module private to sqlite-opfs, returning
the kv-store node entrypoint to its base state.
