+++ yarn-project/kv-store/src/lmdb-v2/index.ts
@@ -1,4 +1,2 @@
 export * from './store.js';
 export * from './factory.js';
-export { storeIdentitySlug } from '../store_identity.js';
-export type { StoreIdentity } from '../store_identity.js';
